Update docs from my experience
The error instructing me to build viewer before running tests was helpful
but I didn't see it covered in the docs so I wasn't sure what was going on.

I also ran into a bunch of snapshot issues, so I updated all them but I'm really not sure that's the right move.